move bottom nav into it's own component
